feat: consolidate services
