+++ models/accounts/response_body.go
@@ -1,7 +1,7 @@
-package tabung_model
+package account_model
 
 type ResponseBody struct {
 	StatusCode   int    `json:"status_code"`
-	Balance      int    `json:"saldo,omitempty"`
+	Balance      *int   `json:"saldo,omitempty"`
 	ErrorMessage string `json:"remark,omitempty"`
 }
